pyproject.toml
[project]
name = "fast_guni_app"
version = "0.0.1"
dependencies = [
    "PyYAML",
    "gunicorn",
    "uvicorn",
    "fastapi",
    "typer",
    "invoke",
    "rich",
    "psutil",
]

[project.scripts]
fgapp = "fast_guni_app.cli:cli"